Organizes into different components
