+++ src/settings.js
@@ -1,0 +1,3 @@
+export default {
+  sidebarWidth: 300
+};
